chore: adapt util functions for new types
- setUniqueCategoriesAndTimes
This function looks for unique timespans within all categories.
If two events overlap they will be added to the same timespan
The start- & endDate will be adapted to be the smallest and largest date

- getRoomTimeTimesCalled
This currently just checks whether a dateSpan fits into a span
and returns the appropriate y position.